added authenticaton

+++ src/main/java/com/yush/journalApp/service/JournalEntryService.java
@@ -1,5 +1,6 @@
 package com.yush.journalApp.service;
 
+import ch.qos.logback.core.encoder.EchoEncoder;
 import com.yush.journalApp.entity.JournalEntry;
 import com.yush.journalApp.entity.User;
 import com.yush.journalApp.repository.JournalEntryRepo;
@@ -27,7 +28,7 @@ public class JournalEntryService {
         journalEntry.setDate(LocalDateTime.now());
         JournalEntry saved = journalEntryRepo.save(journalEntry);
         user.getJournalEntries().add(saved);
-        userService.saveEntry(user);
+        userService.saveUser(user);
         }catch (Exception e){
             System.out.println(e);
             throw new RuntimeException("An error occurred while saving the entry",e);
@@ -47,10 +48,24 @@ public class JournalEntryService {
         return journalEntryRepo.findById(id);
     }
 
-    public void DeleteById(ObjectId id, String userName){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepo.deleteById(id);
+    @Transactional
+    public boolean DeleteById(ObjectId id, String userName){
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if(removed){
+                userService.saveUser(user);
+                journalEntryRepo.deleteById(id);
+            }
+        }catch (Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occurred while deleting the entry", e);
+        }
+        return removed;
     }
+
+//    public List<JournalEntry> findByUserName(String userName){
+//
+//    }
 }
